requirements.txt
Django==2.2.5
djangorestframework==3.10.3
numpy==1.17.2
PyMySQL==0.9.3
pytz==2019.2
sqlparse==0.3.0
